src/test/java/pl/coderstrust/NumbersFromFile/ProcessorTest.java
package pl.coderstrust.NumbersFromFile;

import org.mockito.InjectMocks;
import org.mockito.Mock;

public class ProcessorTest {

    @Mock
    NumbersProcessor numbersProcessor;
    @Mock
    FileProcessor fileProcessor;

    @InjectMocks
    Processor processor;

   /* @Test
    public void shouldProcessProvidedInputFileAndSaveResultToProvidedOutputFile() {
        // given
        when(fileProcessor.readLinesFromFile("fileName.txt")).thenReturn(Arrays.asList("1 2 3", "4 5 6");
        when(numbersProcessor.processLine("1 2 3")).thenReturn("1+2+3=6");

        // when
        processor.process("fileName.txt", "output.txt");

        // then
        verify(fileProcessor).readLinesFromFile(...
        verify(numbersProcessor).processLine(....
        verify(fileProcessor).writeLinesToFile(....
    }*/
}